Release Candidate 1.

+++ library/OpenSkos2/Rdf/ResourceManagerWithSearch.php
@@ -24,15 +24,28 @@ use OpenSkos2\Exception\ResourceAlreadyExistsException;
 use OpenSkos2\Solr\ResourceManager as SolrResourceManager;
 use OpenSkos2\Rdf\ResourceManager;
 use OpenSkos2\Rdf\ResourceCollection;
+use OpenSkos2\Namespaces\SkosXl;
 
 // @TODO Include resource type in insert/delete/search. Now we know it is only concepts
 
 class ResourceManagerWithSearch extends ResourceManager
 {
+
     /**
      * @var \OpenSkos2\Solr\ResourceManager
      */
     protected $solrResourceManager;
+    
+    /**
+     * @param Client $client
+     * @param SolrResourceManager $solrResourceManager
+     */
+    public function __construct(Client $client, SolrResourceManager $solrResourceManager)
+    {
+        parent::__construct($client);
+        $this->solrResourceManager = $solrResourceManager;
+    }
+
 
     /**
      * Use that if inserting a large amount of resources.
@@ -54,16 +67,7 @@ class ResourceManagerWithSearch extends ResourceManager
         $this->solrResourceManager->setIsNoCommitMode($isNoCommitMode);
     }
 
-    /**
-     * @param Client $client
-     * @param SolrResourceManager $solrResourceManager
-     */
-    public function __construct(Client $client, SolrResourceManager $solrResourceManager)
-    {
-        parent::__construct($client);
-        $this->solrResourceManager = $solrResourceManager;
-    }
-
+    
     /**
      * @param \OpenSkos2\Rdf\Resource $resource
      * @throws ResourceAlreadyExistsException
@@ -73,7 +77,7 @@ class ResourceManagerWithSearch extends ResourceManager
         parent::insert($resource);
         $this->solrResourceManager->insert($resource);
     }
-    
+
     /**
      * @param \OpenSkos2\Rdf\ResourceCollection $resourceCollection
      * @throws ResourceAlreadyExistsException
@@ -81,6 +85,16 @@ class ResourceManagerWithSearch extends ResourceManager
     public function insertCollection(ResourceCollection $resourceCollection)
     {
         parent::insertCollection($resourceCollection);
+        $this->solrResourceManager->insertCollection($resourceCollection);
+    }
+
+    /**
+     * @param \OpenSkos2\Rdf\ResourceCollection $resourceCollection
+     * @throws ResourceAlreadyExistsException
+     */
+    public function addToCollection(ResourceCollection $resourceCollection)
+    {
+        parent::addToCollection($resourceCollection);
         $this->solrResourceManager->insertCollection($resourceCollection);
     }
 
@@ -111,9 +125,20 @@ class ResourceManagerWithSearch extends ResourceManager
     public function replace(Resource $resource)
     {
         parent::replace($resource);
-        
-        $this->solrResourceManager->delete($resource);
-        $this->solrResourceManager->insert($resource);
+        if ($resource->getType()->getUri() === \OpenSkos2\Concept::TYPE) {
+            $this->solrResourceManager->delete($resource);
+            $this->solrResourceManager->insert($resource);
+        }
+    }
+
+    /**
+     * @param \OpenSkos2\Rdf\ResourceCollection $resourceCollection
+     * @throws ResourceAlreadyExistsException
+     */
+    public function replaceCollection(ResourceCollection $resourceCollection)
+    {
+        parent::replaceCollection($resourceCollection);
+        $this->solrResourceManager->insertCollection($resourceCollection);
     }
 
     /**
@@ -121,6 +146,14 @@ class ResourceManagerWithSearch extends ResourceManager
      */
     public function delete(Uri $resource)
     {
+        if ($this->resourceType === \OpenSkos2\Concept::TYPE) {
+            $this->client->update("DELETE WHERE {<{$resource->getUri()}> <" . SkosXl::PREFLABEL . "> ?object . "
+                . "?object ?predicate2 ?object2 .}");
+            $this->client->update("DELETE WHERE {<{$resource->getUri()}> <" . SkosXl::ALTLABEL . "> ?object . "
+                . "?object ?predicate2 ?object2 .}");
+            $this->client->update("DELETE WHERE {<{$resource->getUri()}> <" . SkosXl::HIDDENLABEL . "> ?object . "
+                . "?object ?predicate2 ?object2 .}");
+        }
         parent::delete($resource);
         $this->solrResourceManager->delete($resource);
     }
@@ -140,13 +173,13 @@ class ResourceManagerWithSearch extends ResourceManager
     public function search($query, $rows = 20, $start = 0, &$numFound = 0, $sorts = null)
     {
         $filterQueries = null;
-        
+
         if (!empty($this->resourceType)) {
             $filterQueries = [
                 'rdfTypeFilter' => 's_rdfType:"' . $this->resourceType . '"'
             ];
         }
-        
+
         return $this->fetchByUris(
             $this->solrResourceManager->search($query, $rows, $start, $numFound, $sorts, $filterQueries)
         );
